Move validation result classes into their own files.
