+++ src/Stratis.Bitcoin/Consensus/ConsensusManager.cs
@@ -1138,37 +1138,6 @@ namespace Stratis.Bitcoin.Consensus
 
             this.logger.LogTrace("(-)");
         }
-
-        /// <summary>
-        /// Information related to the block full validation process.
-        /// </summary>
-        private class ConnectBlocksResult : ValidationResult
-        {
-            public bool ConsensusTipChanged { get; private set; }
-
-            /// <summary>List of peer IDs to be banned and disconnected.</summary>
-            /// <remarks><c>null</c> in case <see cref="ValidationResult.Succeeded"/> is <c>false</c>.</remarks>
-            public List<int> PeersToBan { get; private set; }
-
-            public ChainedHeader LastValidatedBlockHeader { get; set; }
-
-            public ConnectBlocksResult(bool succeeded, bool consensusTipChanged = true, List<int> peersToBan = null, string banReason = null, int banDurationSeconds = 0)
-            {
-                this.ConsensusTipChanged = consensusTipChanged;
-                this.Succeeded = succeeded;
-                this.PeersToBan = peersToBan;
-                this.BanReason = banReason;
-                this.BanDurationSeconds = banDurationSeconds;
-            }
-
-            public override string ToString()
-            {
-                if (this.Succeeded)
-                    return $"{nameof(this.Succeeded)}={this.Succeeded}";
-
-                return $"{nameof(this.Succeeded)}={this.Succeeded},{nameof(this.ConsensusTipChanged)}={this.ConsensusTipChanged},{nameof(this.PeersToBan)}.{nameof(this.PeersToBan.Count)}={this.PeersToBan.Count},{nameof(this.BanReason)}={this.BanReason},{nameof(this.BanDurationSeconds)}={this.BanDurationSeconds}";
-            }
-        }
     }
 
     /// <summary>
@@ -1184,37 +1153,5 @@ namespace Stratis.Bitcoin.Consensus
     {
         /// <summary>The list of block headers to download.</summary>
         public List<ChainedHeader> BlocksToDownload { get; set; }
-    }
-
-    public class PartialValidationResult : ValidationResult
-    {
-        public ChainedHeaderBlock ChainedHeaderBlock { get; set; }
-
-        /// <inheritdoc/>
-        public override string ToString()
-        {
-            if (this.Succeeded)
-                return base.ToString();
-
-            return base.ToString() + $",{nameof(this.ChainedHeaderBlock)}={this.ChainedHeaderBlock}";
-        }
-    }
-
-    public class ValidationResult
-    {
-        public bool Succeeded { get; set; }
-
-        public int BanDurationSeconds { get; set; }
-
-        public string BanReason { get; set; }
-
-        /// <inheritdoc/>
-        public override string ToString()
-        {
-            if (this.Succeeded)
-                return $"{nameof(this.Succeeded)}={this.Succeeded}";
-
-            return $"{nameof(this.Succeeded)}={this.Succeeded},{nameof(this.BanReason)}={this.BanReason},{nameof(this.BanDurationSeconds)}={this.BanDurationSeconds}";
-        }
     }
 }
