+++ src/Stratis.Bitcoin/Consensus/ValidationResults/PartialValidationResult.cs
@@ -1,0 +1,21 @@
+﻿using Stratis.Bitcoin.Primitives;
+
+namespace Stratis.Bitcoin.Consensus
+{
+    /// <summary>
+    /// Feedback specific to a partial block validation.
+    /// </summary>
+    public sealed class PartialValidationResult : ValidationResult
+    {
+        public ChainedHeaderBlock ChainedHeaderBlock { get; set; }
+
+        /// <inheritdoc/>
+        public override string ToString()
+        {
+            if (this.Succeeded)
+                return base.ToString();
+
+            return base.ToString() + $",{nameof(this.ChainedHeaderBlock)}={this.ChainedHeaderBlock}";
+        }
+    }
+}
